requirements.txt
<<<<<<< HEAD
# Python version requirements
python_version >= "3.10"
python_version < "3.12"

=======
>>>>>>> ccc7477a4320666d819b647b24926a72b08d0bfb
# Web Framework
flask==2.3.3
werkzeug==2.3.7

# AI and Language Processing
langchain==0.1.0
langchain-core==0.1.23
langgraph==0.0.15
google-generativeai==0.3.2
tavily-python==0.2.6
langsmith==0.0.87

# HTTP and API
requests==2.31.0

# Development Tools
python-dotenv==1.0.0
<<<<<<< HEAD
pytest==7.4.3
=======
>>>>>>> ccc7477a4320666d819b647b24926a72b08d0bfb

# Type Hints
typing-extensions==4.9.0

# Utilities
pydantic==1.10.13
tqdm==4.66.1
regex==2023.12.25
 
